Fix the ratings system!

+++ Models/Rating.cs
@@ -2,13 +2,13 @@
 {
 	public class Rating
 	{
-		private int _userRating;
-		public int UserRating { get { return _userRating; } }
+		private double _userRating;
+		public double UserRating { get { return _userRating; } }
 		public User User { get; set; }
 
 		public Movie Movie { get; set; }
 
-		public Rating(int rating, User user, Movie movie)
+		public Rating(double rating, User user, Movie movie)
 		{
 			_userRating = rating;
 			User = user;
